Refactor utils.
