+++ case_studies/galaxy_clustering/data_generation/DES_data_extraction.py
@@ -7,7 +7,7 @@ from astropy import units
 from numpy.core.defchararray import startswith
 from pyvo.dal import sia
 
-from case_studies.galaxy_clustering import cluster_utils as utils
+from case_studies.galaxy_clustering.utils import cluster_utils as utils
 
 DES_DATAPATH = os.environ["HOME"] + "/bliss/case_studies/galaxy_clustering/data/DES_images"
 if not os.path.exists(DES_DATAPATH):
